Dockers

+++ src/eventStore.js
@@ -1,5 +1,11 @@
 #!/usr/bin/env node
 
+/** 
+ * Read config 
+ */
+const env = process.env.NODE_ENV || 'development';
+const config = require(__dirname + '/../config/web-config.json')[env];
+
 /**
  * Module dependencies.
  */
@@ -9,30 +15,30 @@ var http = require('http');
 /**
  * Test data for developement env.
  */
-
-(async () => {
-    await models.sequelize.sync({ force: true, logging: false });
-    if (process.env.MOCK_DATA) {
-        const Storage = require('./storage');
-        let storage = new Storage(models);
-        if (process.env.MOCK_DATA == 2) {
-            var data = require('../test/testData/duplicateEvents.js');
-        } else {
-            var data = require('../test/mockData.js');
-        }
-        await storage.onFinalizedBlock(data.finilizedBlockEvent1);
-        await storage.onFinalizedBlock(data.finilizedBlockEvent2);
-        await storage.onFinalizedBlock(data.finilizedBlockEvent3);
-        await storage.onDeployProcessed(data.deployProcessedEvent1);
-        await storage.onDeployProcessed(data.deployProcessedEvent2);
-        await storage.onDeployProcessed(data.deployProcessedEvent3);
-        await storage.onBlockAdded(data.blockAddedEvent1);
-        await storage.onBlockAdded(data.blockAddedEvent2);
-        await storage.onBlockAdded(data.blockAddedEvent3);
-        console.log('Data loaded!');
-    };
-
-})();
+if (env == 'developement') {
+    (async () => {
+        await models.sequelize.sync({ force: true, logging: false });
+        if (process.env.MOCK_DATA) {
+            const Storage = require('./storage');
+            let storage = new Storage(models);
+            if (process.env.MOCK_DATA == 2) {
+                var data = require('../test/testData/duplicateEvents.js');
+            } else {
+                var data = require('../test/mockData.js');
+            }
+            await storage.onFinalizedBlock(data.finilizedBlockEvent1);
+            await storage.onFinalizedBlock(data.finilizedBlockEvent2);
+            await storage.onFinalizedBlock(data.finilizedBlockEvent3);
+            await storage.onDeployProcessed(data.deployProcessedEvent1);
+            await storage.onDeployProcessed(data.deployProcessedEvent2);
+            await storage.onDeployProcessed(data.deployProcessedEvent3);
+            await storage.onBlockAdded(data.blockAddedEvent1);
+            await storage.onBlockAdded(data.blockAddedEvent2);
+            await storage.onBlockAdded(data.blockAddedEvent3);
+            console.log('Data loaded!');
+        };
+    })();
+}
 
 /**
  * Build the Express app.
@@ -42,50 +48,25 @@ const httpServer = require('./httpServer');
 app = httpServer(models);
 
 /**
- * Get port from environment and store in Express.
+ * Set port.
  */
-
-var port = normalizePort(process.env.PORT || '3000');
-app.set('port', port);
+app.set('port', config.PORT);
 
 /**
  * Create HTTP server.
  */
-
 var server = http.createServer(app);
 
 /**
  * Listen on provided port, on all network interfaces.
  */
-
-server.listen(port);
+server.listen(config.PORT, config.HOST);
 server.on('error', onError);
 server.on('listening', onListening);
 
 /**
- * Normalize a port into a number, string, or false.
- */
-
-function normalizePort(val) {
-    var port = parseInt(val, 10);
-
-    if (isNaN(port)) {
-        // named pipe
-        return val;
-    }
-
-    if (port >= 0) {
-        // port number
-        return port;
-    }
-
-    return false;
-}
-
-/**
  * Event listener for HTTP server "error" event.
  */
-
 function onError(error) {
     if (error.syscall !== 'listen') {
         throw error;
@@ -113,11 +94,10 @@ function onError(error) {
 /**
  * Event listener for HTTP server "listening" event.
  */
-
 function onListening() {
     var addr = server.address();
     var bind = typeof addr === 'string'
         ? 'pipe ' + addr
-        : 'port ' + addr.port;
+        : addr.address + ':' + addr.port;
     console.log('Listening on ' + bind);
 }
